Create Hotel Project 1
This project will continue in Hotel Project 2:

You have been contracted to build a hotel reservation system. Over the next two mini projects you will build this system. This is the system that hotel desk clerks will use to take desk reservations. This system will also generate and save information to a file. This is the instructions for the first half of this project. To be considered complete the program must work without errors.

Assumptions: This is acting as though the only reservations taken are at the desk. All reservations are for one night. The hotel is empty when you start the program.

First, this program will give the full inventory of rooms in the hotel.

Total rooms: 122

There are four types of rooms with these prices:

Standard Rooms, Courtyard - 70 - $125 a night
Standard Room, Scenic - 35 - $145 a night
Deluxe Suite - 15 - $350 a night
Penthouse - 2 - $1135 a night
Second, the program will allow for a room to be reserved for one night. The program will keep track of the inventory as rooms are being reserved.

Third, the program will keep track of revenue brought in for the day from reserved rooms.

Fourth, the user can choose a getTotal() function that will display remaining room inventory, what was reserved and revenue generated for the day.