Create Hotel Project 2
This project is the continuation of Hotel Project 1:

You have been contracted to build a hotel reservation system. Over the next two mini projects you will build this system. This is the system that hotel desk clerks will use to take desk reservations. This system will also generate and save information to a file. This is the instructions for the second half of this project. To be considered complete the program must work without errors.

Assumptions: This is acting as though the only reservations taken are at the desk. All reservations are for one night. The hotel is empty when you start the program.

First, this program will now keep track of the rooms.

Total rooms: 122

There are four types of rooms with these prices:

Standard Rooms, Courtyard - 70 - $125 a night - Rooms 101 thru 170
Standard Room, Scenic - 35 - $145 a night - Rooms 201 thru 235
Deluxe Suite - 15 - $350 a night - Rooms 236 thru 250
Penthouse - 2 - $1135 a night - Rooms 301 and 302
Second, The program will start by asking today's date. I want you to be able to create multiple days.

Third, The program will keep track of who is staying in each room.

Fourth, The program will store the information for each day in a file and be able to retrieve the information from a previous day.

 

PLEASE NOTE: This project must incorporate a class friend or class inheritance structure. Make sure you design it with the fact that the next project will build on this one.